+++ ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: masnus <masnus@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/21 19:54:40 by masnus            #+#    #+#             */
-/*   Updated: 2024/10/23 21:42:01 by masnus           ###   ########.fr       */
+/*   Updated: 2024/10/24 21:57:14 by masnus           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
 	int	i;
